+++ nf/vignat/flowmanager.h
@@ -4,9 +4,9 @@
 #include "lib/flow.h"
 #include "lib/nf_time.h"
 
-int allocate_flowmanager(uint8_t nb_ports,
+int allocate_flowmanager(uint16_t nb_ports,
                          uint16_t starting_port, uint32_t ext_src_ip,
-                         uint8_t ext_device_id,
+                         uint16_t ext_device_id,
                          uint32_t expiration_time,
                          int max_flows);
